wordllm-flask/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-socketio==5.1.1
flask-jwt-extended==4.2.3
python-dotenv==0.19.1
openai==1.39.0
werkzeug==2.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
pydantic==1.10.8
pytz==2023.3
alembic==1.14.1
aiohttp==3.10.11
python-engineio==4.12.0
python-socketio==5.13.0
mammoth==1.9.0
python-docx==1.1.2
pypdf2==3.0.1
lxml==5.4.0
requests==2.32.3
tqdm==4.67.1
qlalchemy==1.4.48
